% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{A dataframe and matrix assembly function.}
\usage{
assemble(
  mod,
  output = "dataframe",
  factor_coding = "mode",
  cond_values = list()
)
}
\arguments{
\item{mod}{Model object to extract data from.}

\item{output}{Desired output format. Defaults to "dataframe".}

\item{factor_coding}{Desired format for categorical variables. Defaults
to "mode" but can be changed to "onehot" for one-hot encoding.}

\item{cond_values}{List of name-value pairs for what values the independent
variables take. Defaults to empty list, which means the function will
use the mode for categorical variables and the mean for numeric variables.}
}
\value{
A dataframe or matrix with the specified values of the independent
variables.
}
\description{
This function allows you to generate a dataframe or matrix with specific
values for the independent variables of a model. Defaults to mode for factor
variables and mean for numeric variables. You can specify values for
specific variables using a list format. You can also specify whether you
want the object returned in a dataframe or matrix format, as well as
whether you want the factor variables to be one-hot encoded or not.
}
\examples{
mod <- lm(mpg ~ hp + cyl, data = mtcars)
assemble(mod)
assemble(mod, output = "matrix")
assemble(mod, factor_coding = "onehot", cond_values = list("hp" = seq(100, 200, 10), "cyl" = 8))
}
