% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_coefs.R
\name{boot_coefs}
\alias{boot_coefs}
\title{A function to bootstrap coefficients from a model and dataset.}
\usage{
boot_coefs(mod, nboots = 500, sample_params = FALSE)
}
\arguments{
\item{mod}{Model object from which to extract data and coefficients.}

\item{nboots}{Number of times to bootstrap the original data set, model,
and coefficients. Defaults to 500.}

\item{sample_params}{Logical value indicating whether to sample random draws
from the multivariate normal distribution implied by the bootstrapped
covariance matrix. Defaults to FALSE.}
}
\value{
A list with the bootstrapped parameters.
}
\description{
This function combines bootstrap resampling of a data set with an estimated
model to generate a sampling distribution for every model coefficient.
It returns a list of the bootstrapped coefficients, their means, and their
covariance matrix. The user can also specify whether they want to sample
parameter vectors from the multivariate normal distribution implied by the
covariance matrix.
}
\examples{
mod <- lm(mpg ~ wt + hp, data = mtcars)
a <- boot_coefs(mod = mod)
str(a)
a$coefs
a$means
a$cov_mat
}
