% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_or_mode.R
\name{mean_or_mode}
\alias{mean_or_mode}
\title{A function to calculate the mean or mode of a variable depending on type.}
\usage{
mean_or_mode(x)
}
\arguments{
\item{x}{A vector of numeric, factor, or character values.}
}
\value{
A scalar indicating the mean or mode of the input vector.
}
\description{
This function calculates the mean or mode of a variable by checking whether
it is numeric, factor or character, or something else. If a user wants to
calculate the mode of an ordinal variable that is coded as an integer, then
the variable must be specified as an ordered factor or be manually given
its character values. The function automatically breaks bimodal variables
by selecting the first mode.
}
\examples{
x <- seq(1, 100, 1)
mean_or_mode(x)
a <- c("Democrat", "Republican", "Independent")
x <- sample(a, 100, replace = T)
table(x)
mean_or_mode(x)

# when two modes are present
a <- c("Democrat", "Democrat", "Independent", "Republican", "Republican")
mean_or_mode(a)

# ordered factors
b <- c(1, 2, 3)
x <- sample(b, 100, replace = T)
mean_or_mode(x)
x[x == 1] <- "Agree"
x[x == 2] <- "Neutral"
x[x == 3] <- "Disagree"
table(x)
mean_or_mode(x)

}
